app/requirements.txt
fastapi==0.115.14
pandas==2.3.0
pandas_gbq==0.29.1
python-dotenv==1.1.1
Requests==2.32.4
uvicorn==0.35.0
gunicorn==23.0.0
sqlmodel==0.0.24
pymysql==1.1.1
passlib==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
pydantic-settings==2.4.0
python-multipart==0.0.20
